PPJB 1.2

+++ aa.sh
@@ -4,15 +4,19 @@ SRCNAME1=$(echo $SRCIPD/Payload/*.app)
 SRCNAME2=${SRCNAME1%.app}
 SRCNAME=${SRCNAME2##*/}
 
-echo aaaa $SRCNAME
-
-
 if [ "$OBJROOT" = "" ] ; then
 	echo Run in xcode. >&2
 	exit 0
 fi
 
+echo source name is $SRCNAME
 cd $OBJROOT
+
+if [ ! -e ../Products/Debug-iphoneos ] ; then
+	echo 'build again ...'
+	exit 0
+fi
+
 cd ../Products/Debug-iphoneos
 
 rm -fr *.dSYM
@@ -20,9 +24,10 @@ rm -fr *.dSYM
 NAME1=$(echo *.app)
 NAME=${NAME1%.app}
 
-echo $NAME
+echo dest name is $NAME
 
 cd $NAME.app
+DESTP=`pwd`
 
 rm -fr _CodeSignature
 rm -fr *.mobileprovision
@@ -35,6 +40,7 @@ for i in $SRCIPD/Payload/$SRCNAME.app/* ; do
 		_CodeSignature)
 			;;
 		Info.plist)
+			plcp $i Info.plist CFBundleIcons CFBundleIcons~ipad
 			;;
 		$SRCNAME)
 			cp $i $NAME
@@ -45,3 +51,11 @@ for i in $SRCIPD/Payload/$SRCNAME.app/* ; do
 			;;
 	esac
 done
+
+if true ; then
+	cp $SRCNAME1/Info.plist /tmp/xipa_src.plist
+	cp $DESTP/Info.plist /tmp/xipa_dst.plist
+	plutil -convert xml1 /tmp/xipa_src.plist
+	plutil -convert xml1 /tmp/xipa_dst.plist
+	echo bcomp /tmp/xipa_src.plist /tmp/xipa_dst.plist
+fi
